requirements.txt
flask
ultralytics
opencv-python
numpy
pillow
torch
torchvision
gdown
